+++ core/input_buffer.hpp
@@ -11,7 +11,7 @@ namespace Aporia
     template<typename T> requires std::is_enum_v<T>
     class InputBuffer
     {
-        static constexpr size_t Size = magic_enum::enum_count<T>();
+        static constexpr u64 Size = magic_enum::enum_count<T>();
         using Buffer = std::bitset<Size>;
 
     public:
@@ -22,31 +22,31 @@ namespace Aporia
 
         void push_state(T code, bool state)
         {
-            size_t index = _index(code);
+            u64 index = _index(code);
             _current_state[index] = state;
         }
 
         bool is_triggered(T code) const
         {
-            size_t index = _index(code);
+            u64 index = _index(code);
             return _is_triggered(index);
         }
 
         bool is_pressed(T code) const
         {
-            size_t index = _index(code);
+            u64 index = _index(code);
             return _is_pressed(index);
         }
 
         bool is_released(T code) const
         {
-            size_t index = _index(code);
+            u64 index = _index(code);
             return _is_released(index);
         }
 
         bool is_any_triggered() const
         {
-            for (size_t index = 0; index < Size; ++index)
+            for (u64 index = 0; index < Size; ++index)
             {
                 if (_is_triggered(index))
                 {
@@ -59,7 +59,7 @@ namespace Aporia
 
         bool is_any_pressed() const
         {
-            for (size_t index = 0; index < Size; ++index)
+            for (u64 index = 0; index < Size; ++index)
             {
                 if (_is_pressed(index))
                 {
@@ -72,7 +72,7 @@ namespace Aporia
 
         bool is_any_released() const
         {
-            for (size_t index = 0; index < Size; ++index)
+            for (u64 index = 0; index < Size; ++index)
             {
                 if (_is_released(index))
                 {
@@ -84,22 +84,22 @@ namespace Aporia
         }
 
     private:
-        size_t _index(T code) const
+        u64 _index(T code) const
         {
             return magic_enum::enum_index(code).value();
         }
 
-        bool _is_triggered(size_t index) const
+        bool _is_triggered(u64 index) const
         {
             return _current_state.test(index) && !_past_state.test(index);
         }
 
-        bool _is_pressed(size_t index) const
+        bool _is_pressed(u64 index) const
         {
             return _current_state.test(index) && _past_state.test(index);
         }
 
-        bool _is_released(size_t index) const
+        bool _is_released(u64 index) const
         {
             return !_current_state.test(index) && _past_state.test(index);
         }
